grpc/requirements.txt
grpcio == 1.71.0
grpcio-tools == 1.71.0
pika == 1.3.2
python-dotenv == 1.1.0